+++ proto/stream_plan.proto
@@ -472,7 +472,7 @@ message StreamFragmentGraph {
     bool is_singleton = 4;
     // Number of table ids (stateful states) for this fragment.
     uint32 table_ids_cnt = 5;
-    // Mark the upstream table ids of this fragment.
+    // Mark the upstream table ids of this fragment, Used for fragments with `Chain`s.
     repeated uint32 upstream_table_ids = 6;
   }
